Rendering changes, hitbox changes and growth changes.
Improved first person rendering.
Dragon camera is now focused around the head rather then the body.
Added a new scaling hitbox to dragons.
The gradual growth has been changed to go slower in the further stages.

+++ src/main/java/by/jackraidenph/dragonsurvival/magic/abilities/Actives/BallLightningAbility.java
@@ -55,7 +55,7 @@ public class BallLightningAbility extends ActiveDragonAbility
 		float f4 = MathHelper.sin(f1);
 		float f5 = MathHelper.cos(f1);
 		
-		float size = DragonStateProvider.getCap(player).map((cap) -> cap.getSize()).get();
+		Double size = DragonStateProvider.getCap(player).map((cap) -> cap.getSize()).get();
 		
 		double x = player.getX() + f4;
 		double y = player.getY() + (size / 20F) - 0.2;
